#59 #60 fix tests for degree and deficiency controllers
